feat: refactor change log entry struct into its own file
